Moving generic components
Moving generic components to root level
